Fixed file permissions

+++ container-commands/build.sh
@@ -2,6 +2,10 @@
 
 echo "Starting build..."
 neos-utils build base
+if [ $? -ne 0 ]; then
+    echo "Base build failed. Aborting..."
+    exit 1
+fi
 
 echo "Linking build..."
 rm -rf "$BUILD_PATH_RELEASE"
